aarch64: fix ANS V wrapper types exposed by complete kernel build
Use the errno API's u64 type, preserve the guarded/locked pointer-map
lookup, pass dynamic buffer data directly to C, and give tos a byte
pointer. These four compatibility fixes preserve storage behavior and
allow the current master kernel to compile with the pinned V1 compiler.
No subsystem is stubbed or compiled out to obtain the build result.

+++ kernel/modules/apple/ans/ans.v
@@ -101,7 +101,7 @@ fn (mut this AnsBlock) write(_handle voidptr, buffer voidptr, loc u64, count u64
 	if result != 0 {
 		C.printf(c'ans: write/flush failed error=%d stage=%u status=0x%x; writes stopped\n',
 			result, C.vinix_ans_stage(), C.vinix_ans_completion_status())
-		errno.set(if result == -12 { errno.erofs } else { errno.eio })
+		errno.set(u64(if result == -12 { errno.erofs } else { errno.eio }))
 		return none
 	}
 	return i64(bytes)
